# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger R2D2",
    "description": "\n\nSome useful links:\n- [The R2D2](https://github.com/Iniciativa-Padawan/r2d2-coruscant)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://r2d2-enyt.onrender.com/"
    }
  ],
  "tags": [
    {
      "name": "movies",
      "description": "Everything about Star Wars movies"
    }
  ],
  "paths": {
    "/movies": {
      "get": {
        "tags": ["movies"],
        "summary": "Finds Movies",
        "description": "Returns a movie list",
        "operationId": "findMovies",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No movies registered"
          }
        }
      },
      "post": {
        "tags": ["movies"],
        "summary": "Add a new movie",
        "description": "Add a new movie",
        "operationId": "addMovie",
        "requestBody": {
          "description": "Create a new movie",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/movies/{movieId}": {
      "put": {
        "tags": ["movies"],
        "summary": "Update an existing movie",
        "description": "Update an existing movie by Id",
        "operationId": "updateMovie",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "description": "Movie id to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update an existent movie",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Movie not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      },
      "delete": {
        "tags": ["movies"],
        "summary": "Deletes a movie",
        "description": "delete a movie",
        "operationId": "deleteMovie",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "description": "Movie id to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid movie value"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Movie": {
        "required": ["name", "trilogy", "sequential", "year"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clh8cwbop0000uxk43kkxcxkb"
          },
          "name": {
            "type": "string",
            "example": "The Phantom Menace"
          },
          "trilogy": {
            "type": "string",
            "example": "sequel | prequel | classic"
          },
          "year": {
            "type": "integer",
            "example": 1999
          },
          "sequential": {
            "type": "integer",
            "example": 1
          }
        },
        "xml": {
          "name": "movie"
        }
      }
    },
    "requestBodies": {
      "Movie": {
        "description": "Movie object that needs to be added",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Movie"
            }
          }
        }
      }
    }
  }
}